+++ src/lib/default-config/eslint.config.mjs
