+++ src/lib/default-config/prettier.config.mjs
