+++ src/lib/default-config/test/data/dangling-commas/index.mjs
