+++ src/lib/default-config/test/data/non-literal-regex/index.mjs
