+++ src/lib/default-config/test/data/windows-style-newline/index.mjs
