+++ src/lib/format-and-lint.mjs
@@ -3,8 +3,8 @@ import { readFile, writeFile } from 'node:fs/promises'
 import { ESLint } from 'eslint'
 import { format as prettierFormat } from 'prettier'
 
-import { eslintConfig as defaultEslintConfig } from './eslint.config'
-import { prettierConfig as defaultPrettierConfig } from './prettier.config'
+import { eslintConfig as defaultEslintConfig } from './default-config/eslint.config'
+import { prettierConfig as defaultPrettierConfig } from './default-config/prettier.config'
 
 const formatAndLint = async ({ eslintConfig = defaultEslintConfig, files, prettierConfig = defaultPrettierConfig, write = false }) => {
   const prettierParseConfig = structuredClone(prettierConfig)
