+++ src/lib/test/data/necessary-semicolon/formatted.txt
