+++ src/lib/test/data/necessary-semicolon/index.mjs
